+++ saleor/graphql/checkout/tests/test_checkout_digital.py
@@ -12,10 +12,12 @@ from ...checkout.mutations import update_checkout_shipping_method_if_invalid
 from ...tests.utils import get_graphql_content
 from .test_checkout import (
     MUTATION_CHECKOUT_CREATE,
-    MUTATION_CHECKOUT_LINES_DELETE,
-    MUTATION_CHECKOUT_LINES_UPDATE,
     MUTATION_CHECKOUT_SHIPPING_ADDRESS_UPDATE,
     MUTATION_UPDATE_SHIPPING_METHOD,
+)
+from .test_checkout_lines import (
+    MUTATION_CHECKOUT_LINES_DELETE,
+    MUTATION_CHECKOUT_LINES_UPDATE,
 )
 
 
@@ -96,8 +98,7 @@ def test_checkout_update_shipping_address(
     """Test updating the shipping address of a digital order throws an error."""
 
     checkout = checkout_with_digital_item
-    checkout_id = graphene.Node.to_global_id("Checkout", checkout.pk)
-    variables = {"checkoutId": checkout_id, "shippingAddress": graphql_address_data}
+    variables = {"token": checkout.token, "shippingAddress": graphql_address_data}
 
     response = api_client.post_graphql(
         MUTATION_CHECKOUT_SHIPPING_ADDRESS_UPDATE, variables
@@ -124,9 +125,8 @@ def test_checkout_update_shipping_method(
     """Test updating the shipping method of a digital order throws an error."""
 
     checkout = checkout_with_digital_item
-    checkout_id = graphene.Node.to_global_id("Checkout", checkout.pk)
     method_id = graphene.Node.to_global_id("ShippingMethod", shipping_method.pk)
-    variables = {"checkoutId": checkout_id, "shippingMethodId": method_id}
+    variables = {"token": checkout.token, "shippingMethodId": method_id}
 
     # Put a shipping address, to ensure it is still handled properly
     checkout.shipping_address = address
@@ -181,10 +181,9 @@ def test_checkout_lines_update_remove_shipping_if_removed_product_with_shipping(
     variant = line.variant
 
     variant_id = graphene.Node.to_global_id("ProductVariant", variant.pk)
-    checkout_id = graphene.Node.to_global_id("Checkout", checkout.pk)
 
     variables = {
-        "checkoutId": checkout_id,
+        "token": checkout.token,
         "lines": [{"variantId": variant_id, "quantity": 0}],
     }
     response = user_api_client.post_graphql(MUTATION_CHECKOUT_LINES_UPDATE, variables)
@@ -210,9 +209,8 @@ def test_checkout_line_delete_remove_shipping_if_removed_product_with_shipping(
     line = checkout.lines.first()
 
     line_id = graphene.Node.to_global_id("CheckoutLine", line.pk)
-    checkout_id = graphene.Node.to_global_id("Checkout", checkout.pk)
 
-    variables = {"checkoutId": checkout_id, "lineId": line_id}
+    variables = {"token": checkout.token, "lineId": line_id}
     response = user_api_client.post_graphql(MUTATION_CHECKOUT_LINES_DELETE, variables)
     content = get_graphql_content(response)